+++ src/sw/Core/src/MainMenuState.h
@@ -19,8 +19,8 @@ class MainMenuState: public State
 		{
 			Adjust
 		};
-		int m_temp_first_room = 22.0;
-		int m_temp_second_room = 23.0;
+		float m_temp_first_room = 22.0f;
+		float m_temp_second_room = 23.0f;
 };
 	
 #endif
